puesta en comun

+++ src/main/java/com/solera/forum2/repository/ThreadRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.solera.forum2.model.ThreadForum;
 
-public interface ThreadRepository extends JpaRepository<ThreadForum, Integer>{
+public interface ThreadRepository extends JpaRepository<ThreadForum, Long>{
 
 }
